+++ site/src/xServices/auth/authXService.ts
@@ -426,7 +426,7 @@ export const authMachine =
             id: "signOut",
             onDone: [
               {
-                actions: ["unassignMe", "clearAuthError"],
+                actions: ["redirectAfterLogout", "unassignMe", "clearAuthError"],
                 target: "gettingMethods",
               },
             ],
@@ -470,23 +470,21 @@ export const authMachine =
           return await API.login(event.email, event.password)
         },
         signOut: async () => {
-          // Get app hostname so we can see if we need to log out of app URLs.
-          // We need to load this before we log out of the API as this is an
-          // authenticated endpoint.
-          const appHost = await API.getApplicationsHost()
-          await API.logout()
-
-          if (appHost.host) {
-            const redirect_uri = encodeURIComponent(window.location.href)
-            // The path doesn't matter but we use /api because the dev server
-            // proxies /api to the backend.
-            const uri = `${window.location.protocol}//${appHost.host.replace(
-              "*",
-              "coder-logout",
-            )}/api/logout?redirect_uri=${redirect_uri}`
-            window.location.replace(uri)
-          }
+          await API.logout
         },
+//         signOut: async () => {
+//           const appHost = await API.getApplicationsHost()
+//           await API.logout
+// 
+//           if (appHost.host) {
+//             const redirect_uri = encodeURIComponent(window.location.href)
+//             const uri = `${window.location.protocol}//${appHost.host.replace(
+//               "*",
+//               "coder-logout",
+//             )}/api/logout?redirect_uri=${redirect_uri}`
+//             window.location.replace(uri)
+//           }
+//         },
         getMe: API.getUser,
         getMethods: API.getAuthMethods,
         updateProfile: async (context, event) => {
